scene repair and prefab stuff

+++ Assets/Scripts/CustomerScene/Customers/OrderEvaluation.cs
@@ -3,7 +3,7 @@ namespace CustomerScene.Customers
     public struct OrderEvaluation
     {
         public bool IsAccepted;
-        public float PricePaid;
+        public int PricePaid;
 
     }
 }
